Removed protocol layer dependency on Libp2pEndpoint
