+++ src/connection/Libp2pEndpoint.js
@@ -4,37 +4,33 @@ const PeerInfo = require('peer-info')
 const pull = require('pull-stream')
 const debug = require('debug')('streamr:connection:libp2pendpoint')
 const { callbackToPromise, getAddress } = require('../util')
+const Endpoint = require('./Endpoint')
 const Libp2pBundle = require('./Libp2pBundle')
 
 const HANDLER = '/streamr/v1/'
-
-const events = Object.freeze({
-    PEER_DISCOVERED: 'streamr:peer:discovery',
-    PEER_CONNECTED: 'streamr:peer:connect',
-    PEER_DISCONNECTED: 'streamr:peer:disconnect',
-    MESSAGE_SENT: 'streamr:message-sent',
-    MESSAGE_RECEIVED: 'streamr:message-received'
-})
 
 class Libp2pEndpoint extends EventEmitter {
     constructor(node) {
         super()
         this.node = node
 
+        this.endpoint = new Endpoint()
+        this.endpoint.implement(this)
+
         node.peerInfo.multiaddrs.forEach((ma) => debug('listening on: %s', ma.toString()))
 
         node.handle(HANDLER, (protocol, conn) => this.onReceive(protocol, conn))
         node.on('peer:discovery', (peer) => {
             debug('peer %s discovered', getAddress(peer))
-            this.emit(events.PEER_DISCOVERED, peer)
+            this.emit(Endpoint.events.PEER_DISCOVERED, peer)
         })
         node.on('peer:connect', (peer) => {
             debug('peer %s connected', getAddress(peer))
-            this.emit(events.PEER_CONNECTED, peer)
+            this.emit(Endpoint.events.PEER_CONNECTED, peer)
         })
         node.on('peer:disconnect', (peer) => {
             debug('peer %s disconnected', getAddress(peer))
-            this.emit(events.PEER_DISCONNECTED, peer)
+            this.emit(Endpoint.events.PEER_DISCONNECTED, peer)
         })
     }
 
@@ -47,7 +43,7 @@ class Libp2pEndpoint extends EventEmitter {
 
             pull(pull.values([message]), conn)
 
-            this.emit(events.MESSAGE_SENT, {
+            this.emit(Endpoint.events.MESSAGE_SENT, {
                 recipient,
                 message
             })
@@ -64,7 +60,7 @@ class Libp2pEndpoint extends EventEmitter {
                 pull.drain((message) => {
                     debug('received from peer %s message with data "%s"', sender instanceof PeerInfo ? getAddress(sender) : '', message)
 
-                    this.emit(events.MESSAGE_RECEIVED, {
+                    this.emit(Endpoint.events.MESSAGE_RECEIVED, {
                         sender,
                         message
                     })
@@ -152,7 +148,6 @@ async function startLibp2pNode(host, port, privateKey, enablePeerDiscovery, boot
 }
 
 module.exports = {
-    events,
     async createEndpoint(host, port, privateKey, enablePeerDiscovery = false, bootstrapNodes) {
         return startLibp2pNode(host, port, privateKey, enablePeerDiscovery, bootstrapNodes).then((n) => new Libp2pEndpoint(n))
     }
